# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dispatcher API",
    "version": "1.0.0",
    "description": "Backend for the Dispatcher Expo app — auth, onboarding, drives board, balances."
  },
  "servers": [{ "url": "https://api-production-f4ac.up.railway.app" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/v1/me": {
      "get": {
        "security": [{ "bearerAuth": [] }],
        "summary": "Current user",
        "responses": { "200": { "description": "OK" } }
      }
    },
...
    "/v1/drives": {
      "get": {
        "security": [{ "bearerAuth": [] }],
        "summary": "List drives. status=open|active|history|assigned|mine; completed=true for public completed feed. Items include poster (and assignee when present). Open board: favorited posters float to top with posterIsFavorite.",
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "security": [{ "bearerAuth": [] }],
        "summary": "Post a drive (title, passenger phone, vehicle class, seats, trip type; optional address + extra info)",
        "responses": { "201": { "description": "Created" } }
      }
    },
...
    "/v1/balances": {
      "get": {
        "security": [{ "bearerAuth": [] }],
        "summary": "List balances for current user",
        "responses": { "200": { "description": "OK" } }
      }
    },
...
    "/v1/favorites": {
      "get": {
        "security": [{ "bearerAuth": [] }],
        "summary": "List users the current user has favorited",
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "security": [{ "bearerAuth": [] }],
        "summary": "Favorite a user (dispatcher↔driver). Body: { userId }",
        "responses": { "201": { "description": "Created" } }
      }
    },
...
  }
}